Add initial Vagrant/Ansible configuration
This isn't an exhaustive installation of all the packages we'll
need, but it's a good start. Does initial setup of PostgreSQL,
PostGIS, Nginx, Scala/SBT, Node/NPM, etc.

This is based off of the nyc-trees setup, which has the most
fleshed out implementation of @hectcastro's ansible-galaxy roles.
One major difference being that we're using Ubuntu Precise (12.04)
in order to be able to run tests using Travis.

Some directories aren't where we typically place them. These can
be shifted around in future tasks, if needed, when setting up the
underlying components.