Add notification delay option and use in geolocation alert
The geolocation API varies widely on how long it takes
to respond. Seems to be relatively quick on safari/firefox
assuming a location was already found, slower on some chromes.

Delaying the notification a bit is a relatively quick solution
to fix the common case where the geolocation very quickly locks
when previously allowed, but not immediately.

Other solutions would be more extensive, such as first making
a getCurrentLocation call with timeout: 0, maximumAge: Infinity
to get a cached position. This could be used to determine
whether the user has allowed geolocation. Once determined,
the actual getCurrentPosition call could be made.