Add nearby amenity search for rerouting
This is basically a thin wrapper around Google's Places search
Javascript API. It doesn't do any rerouting, it just queries Places
based on the user's preferences. To use the functionality, click
"Reroute" when navigating a route and then choose either "Bench" or
"Bathroom". A list of nearby locations which _might_ have the requested
amenity will be outputted to the console.