Closes #110: Add 'resume route' option to navigate controller
Adds a 'destination' object to Navigation service, which
stores current destination as set by the overview controller.

If our current navigation destination is significantly different
from that point, assume we're rerouting to a temp location and display
UI accordingly.

Removed NavigationQueue as the original idea to use it didn't work.

TODO: Additional tests if necessary