Store last navigated route in RouteCache, so walk doesn't reset
Navigating GoForWalk -> Overview -> Navigate -> Overview would generate
a new walk route when going back to Overview for the second time.
This adds a state param and a route cache so we can store the route and
display it at that time instead of generating a new route.

Deletes the MapRoute service, it's utility was marginal, and we needed
the setGeojson function to not be abstracted away.